Carrito de compras
Se añade la ventana del fxml Carrito de compras
Se añade la logica y metodos en las clases Producto y CarritoCompras
Se crea una clase para gestionar productos
Se añaden productos de ejemplo en el txt
Se modifican algunos metodos del carrito y productos en persistencia
Se añade una linea de codigo en iniciar programa (main)

+++ LaboratorioColecciones/src/main/java/co/edu/uniquindio/laboratoriocolecciones/model/Producto.java
@@ -8,14 +8,12 @@ import lombok.*;
 @NoArgsConstructor//constructor vacio
 @ToString// el to string
 @Builder // es aquel que nos permite acceder a estos metodos lombok
+
 public class Producto {
-
-
     private String codigo;
     private String nombreProducto;
-    private Integer precio;
-    private Integer cantidad;
-
-
+    private double precio;
+    private int cantidad;
 
 }
+
